requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
httpx>=0.25.0

# MCP (Model Context Protocol) dependencies
mcp>=1.0.0

# AI and NLP
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
rich>=13.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0